[level 2] Title: 피보나치 수, Time: 2.29 ms, Memory: 70.7 MB -BaekjoonHub

+++ Java/프로그래머스/2/12945. 피보나치 수/피보나치 수.java
@@ -2,14 +2,14 @@ import java.math.*;
 
 class Solution {
     public int solution(int n) {
-        long[] memo = new long[n+1];
+        int[] memo = new int[n+1];
         
-        memo[0] = 0L;
-        memo[1] = 1L;
+        memo[0] = 0;
+        memo[1] = 1;
         
         for (int i=2; i<=n; i++)
-            memo[i] = (memo[i-2] + memo[i-1]) % 1234567L;
+            memo[i] = (memo[i-2] + memo[i-1]) % 1234567;
         
-        return (int) memo[n];
+        return memo[n];
     }
 }
